Change the throughput per dc to be double

+++ src/java/org/apache/cassandra/service/StorageService.java
@@ -1458,13 +1458,13 @@ public class StorageService extends NotificationBroadcasterSupport implements IE
         return DatabaseDescriptor.getStreamThroughputOutboundMegabitsPerSec();
     }
 
-    public void setInterDCStreamThroughputMbPerSec(int value)
+    public void setInterDCStreamThroughputMbPerSec(double value)
     {
         DatabaseDescriptor.setInterDCStreamThroughputOutboundMegabitsPerSec(value);
         logger.info("setinterdcstreamthroughput: throttle set to {}", value);
     }
 
-    public int getInterDCStreamThroughputMbPerSec()
+    public double getInterDCStreamThroughputMbPerSec()
     {
         return DatabaseDescriptor.getInterDCStreamThroughputOutboundMegabitsPerSec();
     }
